# apps/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListDto"
                }
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details with posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithPostsDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
    "/posts": {
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "published",
            "required": false,
            "in": "query",
            "description": "Filter by published status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsListDto"
                }
              }
            }
          }
        },
        "summary": "Get all posts",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithAuthorDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create a new post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post details with author",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithAuthorDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "summary": "Get a post by ID",
        "tags": [
          "posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithAuthorDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "summary": "Update a post",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Post deleted successfully"
          },
          "404": {
            "description": "Post not found"
          }
        },
        "summary": "Delete a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/publish": {
      "patch": {
        "operationId": "PostsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post published successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithAuthorDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "summary": "Publish a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/unpublish": {
      "patch": {
        "operationId": "PostsController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post unpublished successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithAuthorDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "summary": "Unpublish a post",
        "tags": [
          "posts"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS Zod Prisma API",
    "description": "API with auto-generated TypeScript client",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "users",
      "description": ""
    },
    {
      "name": "posts",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "UsersListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "cuid"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "name": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "email",
                "name",
                "createdAt",
                "updatedAt"
              ],
              "title": "User",
              "description": "A user entity"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "title": "UsersList",
        "description": "Paginated list of users"
      },
      "UserWithPostsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "posts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "cuid"
                },
                "title": {
                  "type": "string"
                },
                "published": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "title",
                "published"
              ]
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt",
          "updatedAt",
          "posts"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "User display name"
          }
        },
        "required": [
          "email"
        ],
        "title": "CreateUser",
        "description": "Data required to create a new user"
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt",
          "updatedAt"
        ],
        "title": "User",
        "description": "A user entity"
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "nullable": true,
            "description": "User display name"
          }
        },
        "title": "UpdateUser",
        "description": "Data to update an existing user"
      },
      "PostsListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "cuid"
                },
                "title": {
                  "type": "string"
                },
                "content": {
                  "type": "string",
                  "nullable": true
                },
                "published": {
                  "type": "boolean"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "authorId": {
                  "type": "string",
                  "format": "cuid"
                },
                "author": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "cuid"
                    },
                    "name": {
                      "type": "string",
                      "nullable": true
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "email"
                  ]
                }
              },
              "required": [
                "id",
                "title",
                "content",
                "published",
                "createdAt",
                "updatedAt",
                "authorId",
                "author"
              ]
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "title": "PostsList",
        "description": "Paginated list of posts"
      },
      "PostWithAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "published": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string",
            "format": "cuid"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "cuid"
              },
              "name": {
                "type": "string",
                "nullable": true
              },
              "email": {
                "type": "string",
                "format": "email"
              }
            },
            "required": [
              "id",
              "name",
              "email"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "published",
          "createdAt",
          "updatedAt",
          "authorId",
          "author"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "description": "Post content"
          },
          "published": {
            "default": false,
            "type": "boolean",
            "description": "Publication status"
          },
          "authorId": {
            "type": "string",
            "format": "cuid",
            "description": "Author user ID"
          }
        },
        "required": [
          "title",
          "authorId"
        ],
        "title": "CreatePost",
        "description": "Data required to create a new post"
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "nullable": true,
            "description": "Post content"
          },
          "published": {
            "type": "boolean",
            "description": "Publication status"
          }
        },
        "title": "UpdatePost",
        "description": "Data to update an existing post"
      }
    }
  }
}